pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crowd-detection-system"
version = "1.0.0"
description = "Real-time crowd detection and monitoring system using YOLOv8"
authors = [{name = "Crowd Detection Team"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "ultralytics==8.0.196",
    "opencv-python==4.8.1.78",
    "streamlit==1.28.1",
    "numpy==1.24.3",
    "Pillow==10.0.1",
    "matplotlib==3.7.2",
    "seaborn==0.12.2",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "flake8",
    "mypy",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]
exclude = ["tests*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
